backend/doctomd_service/pyproject.toml
[project]
name = "doctomd-service"
version = "0.1.0"
description = "FastAPI service for converting documents to markdown"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "doctomarkdown>=0.2.0",
    "fastapi>=0.117.1",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "openai>=1.108.1",
    "aiofiles>=24.0.0",
    "pydantic>=2.0.0",
    "pytesseract>=0.3.13",
]
